requirements.txt
colorama
dicttoxml
PyYAML
tqdm
python-magic